auth-session-validator/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# ─── HTTP Client (Active Validation) ─────────────────────────────
httpx==0.27.0
requests==2.31.0

# ─── MITM Proxy (Dynamic Analysis) ───────────────────────────────
mitmproxy==10.3.1

# ─── JWT Analysis & Attacks ──────────────────────────────────────
PyJWT==2.8.0
cryptography==42.0.7

# ─── APK Static Analysis ─────────────────────────────────────────
androguard==4.1.2

# ─── PDF Report Generation ───────────────────────────────────────
reportlab==4.2.0
weasyprint==62.1

# ─── Utilities ───────────────────────────────────────────────────
colorama==0.4.6
rich==13.7.1
python-dotenv==1.0.1
pydantic==2.7.1
# ─── Machine Learning & AI ───────────────────────────────────────
scikit-learn==1.4.2
numpy==1.26.4
google-generativeai==0.5.2
